Defines a new way to specify hot redux form reads from the state
You can pass a new optional property `getFormState(state)` to the
`reduxForm` HOC.  This function takes the current state as an arg
and is expected to return the current form state as a plain JS Object.

This is useful for working with stores use an Immutable object as their
root object, and can also be used as a more flexible version of the 
`reduxMountPoint` property that can read from anywhere in the store.